==> notebooks/pipeline/1.cleaning_convert_types.ipynb <==
import pandas as pd
import os

# 🔹 Chemins
input_path = "../../data/raw/Classeur1.csv"
output_dir = "../../data/cleaning"
output_parquet_path = os.path.join(output_dir, "step1_typed.parquet")

# 🔹 Chargement des données
df = pd.read_csv(input_path, sep=";")

# 🔹 Conversion de 'term'
df['term'] = df['term'].str.extract(r'(\d+)').astype(float)

# 🔹 Conversion de 'emp_length'
emp_length_map = {
    '10+ years': 10.0, '9 years': 9.0, '8 years': 8.0, '7 years': 7.0,
    '6 years': 6.0, '5 years': 5.0, '4 years': 4.0, '3 years': 3.0,
    '2 years': 2.0, '1 year': 1.0, '< 1 year': 0.0, 'n/a': None
}
df['emp_length'] = df['emp_length'].map(emp_length_map)

# 🔹 Conversion des dates
df['issue_d'] = pd.to_datetime(df['issue_d'], errors='coerce')
df['earliest_cr_line'] = pd.to_datetime(df['earliest_cr_line'], errors='coerce')

# Nettoyage du champ (retirer % s'il y en a)
df['int_rate'] = df['int_rate'].astype(str).str.replace('%', '', regex=False)

# Conversion en float
df['int_rate'] = pd.to_numeric(df['int_rate'], errors='coerce')

# Division par 100 pour avoir un taux en décimal
df['int_rate'] = df['int_rate'] / 100

# 🔹 Création du dossier de sortie si besoin
os.makedirs(output_dir, exist_ok=True)

# 🔹 Sauvegarde au format Parquet
df.to_parquet(output_parquet_path, index=False)
print(f"✅ Fichier Parquet sauvegardé : {output_parquet_path}")
